Add path/flow support for Clang Static Analyzer *.plist reports
ClangStaticAnalyzer is good at finding issues which involve a certain
path in code, with given constraints and values of particular variables.
Its HTML reports visualize the full path from the start to the final
location, showing all the intermediate steps and the assumptions taken.
The path and conditions are very often critical to understand why CSA
reports a given issue. Without them one can easily judge CSA for
reporting a false positive, while in fact the issue is there, but the
conditions are non-obvious.

Currently sonar-cxx only reports the final location of a given issue.
This PR addresses this shortcoming and closes the gap to CSA's HTML
reports. With this PR SonarQube is now able to present the full path to
a given issue, which provides similar UX to CSA's HTML reports.

Closes #1707.